Removed the extractor interface
