--- CaptureSnippets/IncludeExtracotrs/IIncludeExtractor.cs
@@ -1,7 +1,0 @@
-namespace CaptureSnippets.IncludeExtracotrs
-{
-    public interface IIncludeExtractor
-    {
-        string Extract(string line);
-    }
-}
