--- CaptureSnippets/IncludeExtracotrs/NoOpUsingExtractor.cs
@@ -1,10 +1,0 @@
-﻿namespace CaptureSnippets.IncludeExtracotrs
-{
-    public class NoOpUsingExtractor : IIncludeExtractor
-    {
-        public string Extract(string line)
-        {
-            return null;
-        }
-    }
-}
