+++ CaptureSnippets/IncludeExtractors/NoOpUsingExtractor.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace CaptureSnippets.IncludeExtractors
+{
+    public class NoOpUsingExtractor
+    {
+        public static Func<string, string> Extract = line => null;
+    }
+}
